Package refactoring. Add mos_tests package for python tests. Add shell_tests for shell tests.
