+++ mos_tests/heat/templates/empty_heat_template.yaml
