+++ mos_tests/heat/templates/empty_heat_template_v2.yaml
