+++ shell_tests/murano/test.sh
