+++ shell_tests/murano/test_murano_cli.sh
