+++ shell_tests/neutron/test_sighup_on_controller.sh
